face_requirements.txt
# Face Recognition Project Requirements
# 설치 방법: pip install -r face_requirements.txt

# Computer Vision
opencv-python>=4.5.0

# Numerical Computing
numpy>=1.19.0

# Image Processing
Pillow>=8.0.0

# Deep Learning Framework
tensorflow>=2.8.0

# Machine Learning Utilities
scikit-learn>=1.0.0

# Data Visualization
matplotlib>=3.3.0

